Repl setup & teardown

+++ repl/Repl.js
@@ -1,13 +1,20 @@
 const pty = require('node-pty');
+const COMMANDS = require('./LangCommands.js');
 
 const Repl = {
-  new(command) {
-    return Object.create(this.init(command));
-  },
+  language: null,
+  process: null,
 
-  init(command) {
-    this.process = pty.spawn(command);
-    return this;
+  init(language) {
+    const command = COMMANDS[language];
+    if (command) {
+      this.process = pty.spawn(command);
+      this.language = language;
+
+      console.log(`INITIALIZED ${command}`);
+      return this;
+    } 
+    return null;
   },
 
   on(event, callback) {
@@ -33,7 +40,8 @@ const Repl = {
   },
 
   kill() {
-    this.process.kill();
+    if (this.process) this.process.kill();
+    this.language = null;
   },
 
   id() {
